chore: fix clippy warnings from rust 1.40 - step 4
