+++ src/data/mod.rs
@@ -7,12 +7,26 @@
 // specific language governing permissions and limitations relating to use of the SAFE Network
 // Software.
 
+pub mod blob;
+#[cfg(test)]
+mod blob_tests;
 pub mod map;
 #[cfg(test)]
 mod map_tests;
+pub mod sequence;
+#[cfg(test)]
+mod sequence_tests;
+pub use blob::{
+    Address as BlobAddress, BlobData, Kind as BlobKind, PrivateBlob, PublicBlob,
+    MAX_BLOB_SIZE_IN_BYTES,
+};
 pub use map::{
     Cmd as MapCmd, DataEntries as MapEntries, DataHistories as MapKeyHistories, MapData,
     MapTransaction, PrivateMap, PrivateSentriedMap, PublicMap, PublicSentriedMap,
     SentriedCmd as SentriedMapCmd, SentryOption, StoredValue as MapValue,
     StoredValues as MapValues,
+};
+pub use sequence::{
+    AppendOperation, DataEntry as SequenceEntry, PrivateSentriedSequence, PrivateSequence,
+    PublicSentriedSequence, PublicSequence, SequenceBase, SequenceData, Values as SequenceValues,
 };
